refactor(leaf): move component, entity, peer etc into separate files
